// src/Umbraco.Community.PropertyVisibility/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'https://localhost:44300' | (string & {});
};

export type HiddenFieldsResponseModel = {
    propertyTypeKeys: Array<string>;
    containerKeys: Array<string>;
    disabled: boolean;
    rootResolution: RootResolutionSource;
    matchedSite?: MatchedSiteModel | null;
    warnings: Array<string>;
};

export type MatchedSiteModel = {
    label: string;
    reason: SiteMatchReason;
};

export type ProblemDetails = {
    type?: string | null;
    title?: string | null;
    status?: number | null;
    detail?: string | null;
    instance?: string | null;
    [key: string]: unknown | string | null | string | null | number | null | string | null | string | null | undefined;
};

export type RootResolutionSource = 'Document' | 'Parent' | 'None' | 'RecycleBin';

export type SiteMatchReason = 'None' | 'Key' | 'Name' | 'Default';

export type GetUmbracoPropertyVisibilityV1HiddenFieldsData = {
    body?: never;
    path?: never;
    query?: {
        documentKey?: string;
        contentTypeKey?: string;
        parentKey?: string;
    };
    url: '/umbraco/property-visibility/v1/HiddenFields';
};

export type GetUmbracoPropertyVisibilityV1HiddenFieldsErrors = {
    /**
     * Bad Request
     */
    400: ProblemDetails;
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetUmbracoPropertyVisibilityV1HiddenFieldsError = GetUmbracoPropertyVisibilityV1HiddenFieldsErrors[keyof GetUmbracoPropertyVisibilityV1HiddenFieldsErrors];

export type GetUmbracoPropertyVisibilityV1HiddenFieldsResponses = {
    /**
     * OK
     */
    200: HiddenFieldsResponseModel;
};

export type GetUmbracoPropertyVisibilityV1HiddenFieldsResponse = GetUmbracoPropertyVisibilityV1HiddenFieldsResponses[keyof GetUmbracoPropertyVisibilityV1HiddenFieldsResponses];